Decrease promise count

+++ nova_vm/src/ecmascript/builtins/control_abstraction_objects/promise_objects/promise_abstract_operations/promise_reaction_records.rs
@@ -70,7 +70,7 @@ pub(crate) enum PromiseReactionHandler<'a> {
     },
     PromiseAll {
         result_promise: Promise<'a>,
-        remaining_unresolved_promise_count: usize,
+        remaining_unresolved_promise_count: u32,
         result_array: Array<'a>,
     },
     Empty,
